clients/mobile-client/requirements.txt
kivy==2.3.0
requests==2.32.3
python-socketio==5.11.0
kivy-deps.sdl2==0.7.0
kivy-deps.glew==0.3.1